Refactor database seeder to create more users and listings

+++ app/Http/Controllers/ListingController.php
@@ -2,14 +2,29 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Middleware\NotSuspended;
 use App\Models\Listing;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Inertia\Inertia;
 
-class ListingController extends Controller
+class ListingController extends Controller implements HasMiddleware
 {
+    public static function middleware()
+    {
+        return [
+            new Middleware(
+                ['auth', 'verified', NotSuspended::class],
+                except: ['index', 'show']
+            )
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -36,6 +51,8 @@ class ListingController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Listing::class);
+
         return Inertia::render('Listing/Create');
     }
 
@@ -44,6 +61,8 @@ class ListingController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Listing::class);
+
         // $newTags = explode(',' ,  $request->tags);
         // $newTags = array_map('trim', $newTags);
         // $newTags = array_filter($newTags);
@@ -75,9 +94,12 @@ class ListingController extends Controller
      */
     public function show(Listing $listing)
     {
+        Gate::authorize('view', $listing);
+
         return Inertia::render('Listing/Show', [
             'listing' => $listing,
-            'user' => $listing->user->only(['name', 'id'])
+            'user' => $listing->user->only(['name', 'id']),
+            'canModify' => Auth::user() ? Auth::user()->can('modify', $listing) : false
         ]);
     }
 
@@ -86,6 +108,8 @@ class ListingController extends Controller
      */
     public function edit(Listing $listing)
     {
+        Gate::authorize('modify', $listing);
+
         return Inertia::render('Listing/Edit', [
             'listing' => $listing
         ]);
@@ -96,6 +120,8 @@ class ListingController extends Controller
      */
     public function update(Request $request, Listing $listing)
     {
+        Gate::authorize('modify', $listing);
+
         $fields = $request->validate([
             'title' => ['required', 'max:255'],
             'desc' => ['required'],
@@ -116,9 +142,9 @@ class ListingController extends Controller
 
         $fields['tags'] = implode(',', array_unique(array_filter(array_map('trim', explode(',',  $request->tags)))));
 
-        $listing->update($fields);
+        $listing->update([...$fields, 'approved' => false]);
 
-        return redirect()->route('home')->with('status', 'Listing updated successfully.');
+        return redirect()->route('dashboard')->with('status', 'Listing updated successfully.');
     }
 
     /**
@@ -126,12 +152,14 @@ class ListingController extends Controller
      */
     public function destroy(Listing $listing)
     {
+        Gate::authorize('modify', $listing);
+
         if ($listing->image) {
             Storage::disk('public')->delete($listing->image);
         }
 
         $listing->delete();
 
-        return redirect()->route('home')->with('status', 'Listing deleted successfully.');
+        return redirect()->route('dashboard')->with('status', 'Listing deleted successfully.');
     }
 }
